Issue #5 - Update SystemC and UVM/SystemC Wrappers to Permit MSVC /W4.
Removed code quality items from project.

Added code quality items to UVM property sheet (i.e /W4 and /Wx).

NOTE: Microsoft observes that the STL is not /Wall safe, so backed down to /W4.

NOTE: MSVC warning 4127 is disabled, as it complains about constant conditions in
control code.  This obviates the need for "constexpr" in if statements which are forced
to true or false.